Fix peerId datatype and disable non-functional cluster tests

+++ http/src/main/java/io/metaloom/qdrant/client/http/model/cluster/operation/AbortTransferOperation.java
@@ -1,29 +1,31 @@
 package io.metaloom.qdrant.client.http.model.cluster.operation;
 
+import java.math.BigInteger;
+
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class AbortTransferOperation extends AbstractShardOperation {
 
 	@JsonProperty("to_peer_id")
-	private long toPeerId;
+	private BigInteger toPeerId;
 
 	@JsonProperty("from_peer_id")
-	private long fromPeerId;
+	private BigInteger fromPeerId;
 
-	public long getToPeerId() {
+	public BigInteger getToPeerId() {
 		return toPeerId;
 	}
 
-	public AbortTransferOperation setToPeerId(long toPeerId) {
+	public AbortTransferOperation setToPeerId(BigInteger toPeerId) {
 		this.toPeerId = toPeerId;
 		return this;
 	}
 
-	public long getFromPeerId() {
+	public BigInteger getFromPeerId() {
 		return fromPeerId;
 	}
 
-	public AbortTransferOperation setFromPeerId(long fromPeerId) {
+	public AbortTransferOperation setFromPeerId(BigInteger fromPeerId) {
 		this.fromPeerId = fromPeerId;
 		return this;
 	}
