+++ http/src/main/java/io/metaloom/qdrant/client/http/model/cluster/operation/DropReplicaOperation.java
@@ -1,17 +1,19 @@
 package io.metaloom.qdrant.client.http.model.cluster.operation;
 
+import java.math.BigInteger;
+
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class DropReplicaOperation extends AbstractShardOperation {
 
 	@JsonProperty("peer_id")
-	private long peerId;
+	private BigInteger peerId;
 
-	public long getPeerId() {
+	public BigInteger getPeerId() {
 		return peerId;
 	}
 
-	public DropReplicaOperation setPeerId(long peerId) {
+	public DropReplicaOperation setPeerId(BigInteger peerId) {
 		this.peerId = peerId;
 		return this;
 	}
